+++ src/main/java/org/openbase/bco/bcozy/controller/group/AuthorizationGroupUsersController.java
@@ -2,13 +2,13 @@ package org.openbase.bco.bcozy.controller.group;
 
 import de.jensd.fx.glyphs.fontawesome.FontAwesomeIcon;
 import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.ReadOnlyStringProperty;
 import javafx.beans.property.SimpleObjectProperty;
-import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.*;
+import javafx.scene.control.ComboBox;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
 import javafx.scene.paint.Color;
 import javafx.util.Duration;
 import javafx.util.StringConverter;
@@ -27,8 +27,6 @@ import org.openbase.jul.exception.printer.ExceptionPrinter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import rst.domotic.unit.UnitConfigType;
-
-import java.util.Optional;
 
 /**
  * @author vdasilva
@@ -135,16 +133,6 @@ public class AuthorizationGroupUsersController {
 
     }
 
-    private boolean askForRemoveConfirmation(UnitConfigType.UnitConfig group, UserViewModel user) {
-        String message = LanguageSelection.getLocalized("removeUserFromGroup.confirmation",
-                user.getName(), group.getLabel());
-
-        Optional<ButtonType> selection = new Alert(Alert.AlertType.CONFIRMATION, message).showAndWait();
-
-        return selection.isPresent() && selection.get().equals(ButtonType.OK);
-
-    }
-
     @FXML
     private void addUser() {
         UnitConfigType.UnitConfig group = selectedGroup.get();
@@ -209,32 +197,6 @@ public class AuthorizationGroupUsersController {
 
         userTable.setItems(usersInGroup);
         this.availableUsersComboBox.setItems(availableUsers);
-    }
-
-    public static class UserViewModel {
-        private final ReadOnlyStringProperty id;
-        private final ReadOnlyStringProperty name;
-
-        public UserViewModel(UnitConfigType.UnitConfig unitConfig) {
-            this.id = new SimpleStringProperty(unitConfig.getId());
-            this.name = new SimpleStringProperty(unitConfig.getUserConfig().getUserName());
-        }
-
-        public String getId() {
-            return id.get();
-        }
-
-        public ReadOnlyStringProperty idProperty() {
-            return id;
-        }
-
-        public String getName() {
-            return name.get();
-        }
-
-        public ReadOnlyStringProperty nameProperty() {
-            return name;
-        }
     }
 
 }
